darshan-changes
•	WWI has its own e-Commerce Platform where they sell WWI Brand Merchandise- full stop missing
•	This WWI infrastructure is on on-premise -full stop missing
 confidence rating helps "to" you estimate the 
relaible(reliable)
recommends to performs(perform) assessment after 24 hours
Congratulations ! You are now ready with your Database Assessment, Please move for the Migration of same.
Congratulations! You are now ready with your Database Assessment, please move for the Migration of the same.
OR
Congratulations! You are now ready with your Database Assessment. Please move for the Migration of the same.